requirements.txt
asgiref==3.4.1
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.10.0
install==1.3.5
Markdown==3.3.6
numpy==1.22.0
pandas==1.3.5
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
zipp==3.7.0
